+++ mesh/src/main/java/io/particle/mesh/bluetooth/connecting/ConnectionState.kt
@@ -1,7 +1,7 @@
 package io.particle.mesh.bluetooth.connecting
 
 import android.bluetooth.BluetoothProfile
-import io.particle.mesh.common.buildIntValueMap
+import io.particle.android.sdk.utils.buildIntValueMap
 
 
 enum class ConnectionState(val intValue: Int)  {