+++ mesh/src/main/java/io/particle/mesh/setup/flow/Flow.kt
@@ -5,6 +5,7 @@ import io.particle.android.sdk.cloud.ParticleDevice.ParticleDeviceType
 import io.particle.android.sdk.cloud.ParticleEventVisibility
 import io.particle.firmwareprotos.ctrl.Network.InterfaceEntry
 import io.particle.firmwareprotos.ctrl.Network.InterfaceType
+import io.particle.mesh.R
 import io.particle.mesh.common.Result
 import io.particle.mesh.setup.flow.modules.bleconnection.BLEConnectionModule
 import io.particle.mesh.setup.flow.modules.cloudconnection.CloudConnectionModule
@@ -13,33 +14,25 @@ import io.particle.mesh.setup.flow.modules.device.NetworkSetupType
 import io.particle.mesh.setup.flow.modules.meshsetup.MeshNetworkToJoin.CreateNewNetwork
 import io.particle.mesh.setup.flow.modules.meshsetup.MeshNetworkToJoin.SelectedNetwork
 import io.particle.mesh.setup.flow.modules.meshsetup.MeshSetupModule
-import io.particle.mesh.R
 import kotlinx.coroutines.delay
 import mu.KotlinLogging
 import java.util.concurrent.TimeUnit
 
 
-// FIXME: put this elsewhere
-enum class MeshDeviceType(val particleDeviceType: ParticleDeviceType) {
-
-    ARGON(ParticleDeviceType.ARGON),
-    BORON(ParticleDeviceType.BORON),
-    XENON(ParticleDeviceType.XENON);
-
-//    companion object {
-//
-//
-//    }
+enum class Gen3ConnectivityType {
+    WIFI,
+    CELLULAR,
+    MESH_ONLY
 }
 
-fun ParticleDeviceType.toMeshDeviceType(): MeshDeviceType {
+fun ParticleDeviceType.toMeshDeviceType(): Gen3ConnectivityType {
     return when (this) {
         ParticleDeviceType.ARGON,
-        ParticleDeviceType.A_SERIES -> MeshDeviceType.ARGON
+        ParticleDeviceType.A_SERIES -> Gen3ConnectivityType.WIFI
         ParticleDeviceType.BORON,
-        ParticleDeviceType.B_SERIES -> MeshDeviceType.BORON
+        ParticleDeviceType.B_SERIES -> Gen3ConnectivityType.CELLULAR
         ParticleDeviceType.XENON,
-        ParticleDeviceType.X_SERIES -> MeshDeviceType.XENON
+        ParticleDeviceType.X_SERIES -> Gen3ConnectivityType.MESH_ONLY
         else -> throw IllegalArgumentException("Not a mesh device: $this")
     }
 }
@@ -70,10 +63,14 @@ class Flow(
         if (hasEthernet) {
             doEthernetSubflow()
         } else {
-            when (flowManager.targetDeviceType) {
-                MeshDeviceType.ARGON -> doArgonFlow()
-                MeshDeviceType.BORON -> doBoronFlow()
-                MeshDeviceType.XENON -> doJoinerSubflow()
+            if (flowManager.targetDeviceType != Gen3ConnectivityType.MESH_ONLY) {
+                meshSetupModule.showNewNetworkOptionInScanner = true
+            }
+            val flowType = getFlowType()
+            when (flowType) {
+                Gen3ConnectivityType.WIFI -> doArgonFlow()
+                Gen3ConnectivityType.CELLULAR -> doBoronFlow()
+                Gen3ConnectivityType.MESH_ONLY -> doJoinerSubflow()
             }
         }
 
@@ -83,6 +80,20 @@ class Flow(
             ParticleEventVisibility.PRIVATE,
             TimeUnit.HOURS.toSeconds(1).toInt()
         )
+    }
+
+    private suspend fun getFlowType(): Gen3ConnectivityType {
+        if (flowManager.targetDeviceType == Gen3ConnectivityType.MESH_ONLY) {
+            return Gen3ConnectivityType.MESH_ONLY
+        }
+
+        deviceModule.ensureNetworkSetupTypeCaptured()
+        return if (deviceModule.networkSetupTypeLD.value!! == NetworkSetupType.JOINER) {
+            Gen3ConnectivityType.MESH_ONLY
+        } else {
+            meshSetupModule.showNewNetworkOptionInScanner = true
+            flowManager.targetDeviceType
+        }
     }
 
     suspend fun runMeshFlowForGatewayDevice() {
@@ -183,7 +194,7 @@ class Flow(
         // gather initial data, perform upfront checks
         deviceModule.ensureDeviceIsUsingEligibleFirmware(
             bleConnModule.targetDeviceTransceiverLD.value!!,
-            flowManager.targetDeviceType
+            flowManager.targetPlatformDeviceType
         )
 
         deviceModule.ensureEthernetDetectionSet()