+++ mesh/src/main/java/io/particle/mesh/setup/ui/GetReadyForSetupFragment.kt
@@ -14,10 +14,11 @@ import androidx.annotation.StringRes
 import androidx.core.view.isVisible
 import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.LifecycleOwner
-import androidx.navigation.fragment.findNavController
 import com.squareup.phrase.Phrase
 import io.particle.android.common.buildRawResourceUri
-import io.particle.mesh.setup.flow.MeshDeviceType
+import io.particle.mesh.R
+import io.particle.mesh.setup.flow.Gen3ConnectivityType
+import io.particle.mesh.setup.isSomSerial
 import io.particle.mesh.setup.ui.HelpTextConfig.ARGON
 import io.particle.mesh.setup.ui.HelpTextConfig.A_SERIES
 import io.particle.mesh.setup.ui.HelpTextConfig.BORON_3G
@@ -26,7 +27,6 @@ import io.particle.mesh.setup.ui.HelpTextConfig.B_SERIES
 import io.particle.mesh.setup.ui.HelpTextConfig.FEATHERWING
 import io.particle.mesh.setup.ui.HelpTextConfig.XENON
 import io.particle.mesh.setup.ui.HelpTextConfig.X_SERIES
-import io.particle.mesh.R
 import kotlinx.android.synthetic.main.fragment_get_ready_for_setup.*
 
 
@@ -66,19 +66,18 @@ class GetReadyForSetupFragment : BaseMeshSetupFragment() {
             HelpTextConfig.FEATHERWING
         } else {
             val barcodeLD = flowManagerVM.flowManager!!.bleConnectionModule.targetDeviceBarcodeLD
-            val isSomSerial = barcodeLD.value?.serialNumber?.toLowerCase()?.startsWith("p00") ?: false
+            val isSomSerial = barcodeLD.value?.serialNumber?.isSomSerial() ?: false
 
             when (flowManagerVM.flowManager!!.targetDeviceType) {
-                MeshDeviceType.ARGON -> { if (isSomSerial) HelpTextConfig.A_SERIES else HelpTextConfig.ARGON }
-                MeshDeviceType.XENON -> { if (isSomSerial) HelpTextConfig.X_SERIES else HelpTextConfig.XENON }
-                MeshDeviceType.BORON -> { if (isSomSerial) HelpTextConfig.B_SERIES else HelpTextConfig.BORON_3G }
+                Gen3ConnectivityType.WIFI -> { if (isSomSerial) HelpTextConfig.A_SERIES else HelpTextConfig.ARGON }
+                Gen3ConnectivityType.MESH_ONLY -> { if (isSomSerial) HelpTextConfig.X_SERIES else HelpTextConfig.XENON }
+                Gen3ConnectivityType.CELLULAR -> { if (isSomSerial) HelpTextConfig.B_SERIES else HelpTextConfig.BORON_3G }
             }
         }
 
         p_getreadyforsetup_antenna_confirmation_speedbump.isVisible = when (config) {
             FEATHERWING,
-            XENON,
-            X_SERIES -> {
+            XENON -> {
                 action_next.isEnabled = true
                 false
             }
@@ -86,7 +85,8 @@ class GetReadyForSetupFragment : BaseMeshSetupFragment() {
             BORON_LTE,
             BORON_3G,
             A_SERIES,
-            B_SERIES -> {
+            B_SERIES,
+            X_SERIES -> {
                 action_next.isEnabled = p_getreadyforsetup_antenna_confirmation_speedbump.isChecked
                 true
             }